pyproject.toml
[tool.poetry]
name = "function-action-oidc"
version = "1.0.0"
description = ""
authors = ["Håkon Treider <hakon.treider@cognite.com>"]

[tool.black]
line-length = 120
target_version = ['py38']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # this makes sort all Python files

[tool.poetry.dependencies]
python = "^3.8"
cognite-sdk-experimental = "^0.59"
pydantic = "^1.8"
python-crontab = "^2.5"
PyYAML = "^5.4"
retry = "^0.9"
humanize = "^3.11"

[tool.poetry.dev-dependencies]
bandit = "^1.7"
black = "21.7b0"
isort = "^5.9"
mypy = "0.910"
pytest = "^6.2"
pytest-cov = "^2.12"
